requirements.txt
pyperclip>=1.8.0
qrcode>=7.0
Pillow>=9.0.0